// app-ui/src/app/backend/model/mapFolderCollection.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MapFolderCollectionBreadcrumbs } from './mapFolderCollectionBreadcrumbs';
import { MapFolderModel } from './mapFolderModel';


export interface MapFolderCollection { 
    readonly items?: Array<MapFolderModel>;
    readonly breadcrumbs?: Array<MapFolderCollectionBreadcrumbs>;
}