// app-ui/src/app/backend/model/mapFolderModel.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MapFolderModelSubFolders } from './mapFolderModelSubFolders';
import { MapFolderModelAccessRights } from './mapFolderModelAccessRights';
import { MapFolderModelMaps } from './mapFolderModelMaps';


export interface MapFolderModel { 
    mapFolderId?: number;
    parentMapFolderId?: number;
    name: string;
    accessRights?: MapFolderModelAccessRights;
    readonly maps?: Array<MapFolderModelMaps>;
    readonly subFolders?: Array<MapFolderModelSubFolders>;
    readonly breadcrumbs?: Array<MapFolderModelSubFolders>;
}